+++ core/animatedSprite.h
@@ -5,7 +5,7 @@
 namespace core {
 	class AnimatedSprite : Sprite {
 	private:
-		int secondsPerFrame = 60;
+		float secondsPerFrame = 60.0f;
 		bool isPlaying = false;
 		
 		int totalFrames;
@@ -18,13 +18,13 @@ namespace core {
 
 		void NextFrame();
 	public:
-		AnimatedSprite(Texture *texture, int frameWidth, int frameHeight, int secondsPerFrame, int scale = 1);
+		AnimatedSprite(Texture *texture, int frameWidth, int frameHeight, float secondsPerFrame, int scale = 1);
 		void Update(float deltaTime);
 		void UpdateAndDraw(float deltaTime, Vector2 position);
 		void Play();
 		void Reset();
 		void Stop();
 
-		static AnimatedSprite * Create(const std::string &textureName, int frameWidth, int frameHeight, int secondsPerFrame, int scale = 1);
+		static AnimatedSprite * Create(const std::string &textureName, int frameWidth, int frameHeight, float secondsPerFrame, int scale = 1);
 	};
 }
